fix(zsh): Fixed httpj to not blow up if no json body was created in editor to POST

+++ zsh/.config/zsh/functions/httpie/httpj.zsh
@@ -3,6 +3,11 @@
 function httpj(){
   FILE_NAME=$(rand 40)
   TMPFILE=$(mktemp /tmp/$FILE_NAME.json)
+  NOT_EMPTY=$(cat TMPFILE | wc -c)
+  if [ $NOT_EMPTY = 0 ]; then
+    echo "content empty. exiting..." 
+    exit 0
+  fi
   nvim $TMPFILE
   http "$@" < $TMPFILE
   rm $TMPFILE
